requirements.txt
bson
Flask
PyMongo
arrow
awsebcli
requests
pycrypto